Import seteam local changes for demo
The original base module had a great deal of code replication and unclean
mixing of data and logic. This commit cleans up a good deal of that, but
it is by no means complete. These changes are primarily asthetic. What
still needs to be done is a slight refactoring to remove the need for
having multiple, largely identical classes for deploying the forwarder
on different platforms.