Update solaris package name 9 to 10
With the newest Splunk release the Solaris package(s) are named with
"...solaris-10-..." instead of the olde "...solaris-9-...". This commit
updates the parameters file to reflect that.

If using old packages, it may now be necessary to explicitely call
out the package source for Solaris. Please use current version(s) of
Splunk.