types: Fix purging when section contains '//'
A change made in 0fe96f7 to support section headings of the form:

```
foo://bar
```

broke purging, where the underlying inifile type generates resources
with titles like:

```
section/setting
```

When a resource is generated with a title 'foo://bar/setting', the
setting name is not parsed.  This is the case with inputs which often
take on the form:

```
[monitor:///var/log/messages]
setting1 = value1
setting2 = value2
```

The result is an error message: "Error:
/Stage[main]/Splunk::Params/Splunk_config[splunk]: Failed to generate
additional resources using 'generate': Got nil value for setting."  See:
issue #69.

This change looks for additional slashes beyond the '://' and will
parse out the following string as the setting name.

This is not very satisfying because it could break a theoretical section
name containing extra slashes, but I have not come across that before,
and the `[monitor:///../../../]` case is **very** common.