Use package_source when not staging packages
Prior to this, if a custom forwarder or splunk package source was
supplied to the package_source parameter and that package did not
need to be staged (because we're targetting an MSI on a UNC path
for example), the package installation would fail with a message
like:

  source is required when using package provider 'windows'

The reason it failed is because the $package_source variable was not
being used as the source of the package. Instead, the
$pkg_source variable is used. $pkg_source is only created if
the file needs to be staged; which in the case of using Windows and a
UNC package source, it does not.

This fixes the problem by using the pick() function to choose either
$pkg_source or $package_source (in that order) as the source of the
package.