Fix .conf file modes for Windows forwarders
Prior to this, the file mode for inputs.conf, outputs.conf, and web.conf
wouuld change from `0664` to `0644` on every Puppet run.

On Windows, we don't want to set the file mode for these files as
handling the permissions on these would best be left for the
puppetlabs/acl module.