Run license_splunkforwarder exec resource whenever ftr file exists.
ftr stands for first time run.

The ftr file is an improved trigger for determining when to run the
license_splunkforwarder exec resource. The ftr file always exists
following an install of splunk. The install could be a first time
install or an upgrade. The ftr file is removed when the license
agreement has been accepted.

Triggering the exec resource based on the presence of the
etc/auth/server.pem prevents the exec resource from being called after
splunk has been upgraded, unless we're willing to delete the server.pem
file.