Partially revert metadata.json
GitHub seems to build the test matrix using the latest version of
metadata.json in the default branch, and not the one in the current
branch.

The previous commit has been rebased on top of master, and this dummy
commit was added to re-introduce unsupported combinations and validate
the behaviour of the test matrix.