Add extract_command as a parameter to splunk::addon
https://dev.splunk.com/enterprise/docs/releaseapps/packageapps/ - "An app package is a compressed tar archive, or tarball, with a .spl, .tgz, or .tar.gz file extension"
The archive module (which this module calls) does not know what a '.spl' file is, and will refuse to untar it.  But it's simply a tarball.

.spl, as an extension, is not unique to Splunk (it means 'spool' in some contexts) so it's probably not smart to edit the archive module to 'know' about .spl, just to satisfy one vendor.

This commit is intended to allow you to keep a .spl source file, and for puppet know what to do with it.

The alternate to rename the file to .tgz is possible, but somewhat non-obvious for users, who may want to keep their spl filename.

It would be possible to put some logic in splunk::addon to detect a .spl source via regexp and pass along an 'untar' extract_command, but because of OS variations I think it's best not to gum up the works with a Linux bias and to leave the work up to a profile that uses splunk.