Allow the user to set encrypted values directly
Before it was only possible to set an encrypted value by setting it in plain text.
If the encrypted value was given, there was and endless re-apply happening
because Splunk encrypts the values during boot time thus changing the value.

Now it's also possible to set the values by providing the encrypted value directly.