Fix upgrading Splunk Enterprise
When upgrading, there was an issue with accepting the license of the new version.
Thus symptoms like described in #339 happened. To ensure we have a clean upgrade
we need to re-enable boot-start after the upgrade.

This will also apply the correct systemd service file when upgrading, before it
was just the service file of the originally installed version. It never changed.